Move tracing tests to new files
