+++ sdk/ai/azure-ai-inference/tests/gen_ai_trace_verifier.py
@@ -14,30 +14,40 @@ class GenAiTraceVerifier:
         attribute_dict = dict(attributes)
 
         for attribute_name in span.attributes.keys():
+
             # Check if the attribute name exists in the input attributes
             if attribute_name not in attribute_dict:
+                print(f"Attribute name {attribute_name} not in dictionary. Return False.")
                 return False
 
             attribute_value = attribute_dict[attribute_name]
+            print(f"Attribute name: `{attribute_name}`. Expected attribute value: `{attribute_value}`.")
+
             if isinstance(attribute_value, list):
                 # Check if the attribute value in the span matches the provided list
                 if span.attributes[attribute_name] != attribute_value:
+                    print(f"Case 1: Attribute values do not match (`{span.attributes[attribute_name]}` != `{attribute_value}`). Return False.")
                     return False
             elif isinstance(attribute_value, tuple):
                 # Check if the attribute value in the span matches the provided list
                 if span.attributes[attribute_name] != attribute_value:
+                    print(f"Case 2: Attribute values do not match (`{span.attributes[attribute_name]}` != `{attribute_value}`). Return False.")
                     return False
             else:
                 # Check if the attribute value matches the provided value
                 if attribute_value == "+":
                     if not isinstance(span.attributes[attribute_name], numbers.Number):
+                        print(f"Actual attribute value is not a number.")
                         return False
                     if span.attributes[attribute_name] < 0:
+                        print(f"Actual attribute value is negative.")
                         return False
                 elif attribute_value != "" and span.attributes[attribute_name] != attribute_value:
+                    print(f"Case 3: Attribute values do not match (`{span.attributes[attribute_name]}` != `{attribute_value}`). Return False.")
                     return False
                 # Check if the attribute value in the span is not empty when the provided value is ""
                 elif attribute_value == "" and not span.attributes[attribute_name]:
+                    print(f"Case 4. Return False.")
                     return False
 
         return True
