+++ sdk/ai/azure-ai-inference/tests/model_inference_test_base.py
@@ -39,9 +39,9 @@ if LOGGING_ENABLED:
 ServicePreparerChatCompletions = functools.partial(
     EnvironmentVariableLoader,
     "azure_ai_chat",
-    azure_ai_chat_endpoint="https://your-deployment-name.eastus2.inference.ai.azure.com",
+    azure_ai_chat_endpoint="https://your-deployment-name.eastus2.models.ai.azure.com",
     azure_ai_chat_key="00000000000000000000000000000000",
-    azure_ai_chat_model="model-name"
+    azure_ai_chat_model="mistral-large-2411"
 )
 
 #
@@ -56,7 +56,7 @@ ServicePreparerChatCompletions = functools.partial(
 ServicePreparerAOAIChatCompletions = functools.partial(
     EnvironmentVariableLoader,
     "azure_openai_chat",
-    azure_openai_chat_endpoint="https://your-deployment-name.openai.azure.com/openai/deployments/gpt-4o",
+    azure_openai_chat_endpoint="https://your-deployment-name.openai.azure.com/openai/deployments/gpt-4o-deployment",
     azure_openai_chat_key="00000000000000000000000000000000",
     azure_openai_chat_api_version="yyyy-mm-dd-preview"
 )
@@ -70,7 +70,7 @@ ServicePreparerAOAIChatCompletions = functools.partial(
 ServicePreparerEmbeddings = functools.partial(
     EnvironmentVariableLoader,
     "azure_ai_embeddings",
-    azure_ai_embeddings_endpoint="https://your-deployment-name.eastus2.inference.ai.azure.com",
+    azure_ai_embeddings_endpoint="https://your-deployment-name.eastus2.models.ai.azure.com",
     azure_ai_embeddings_key="00000000000000000000000000000000",
 )
